refactor: simplify tempo session transport internals
