+++ src/client/Methods.ts
@@ -2,6 +2,7 @@ export { evm } from '../evm/client/index.js'
 export { stripe } from '../stripe/client/index.js'
 export { subscription } from '../tempo/client/Subscription.js'
 export { tempo } from '../tempo/client/index.js'
-export { session } from '../tempo/session/client/Session.js'
+export { session as sessionMethod } from '../tempo/session/client/Session.js'
+export { sessionManager as session } from '../tempo/session/client/SessionManager.js'
 export { session as sessionLegacy } from '../tempo/legacy/client/index.js'
 export { sessionManager as sessionLegacyManager } from '../tempo/legacy/client/index.js'
