+++ src/client/Mppx.test-d.ts
@@ -10,10 +10,16 @@ import type * as AutoSwap from '../tempo/internal/auto-swap.js'
 import * as Methods from '../tempo/Methods.js'
 import * as z from '../zod.js'
 import * as Fetch from './internal/Fetch.js'
+import { session, sessionMethod } from './Methods.js'
 import * as Mppx from './Mppx.js'
 import * as Transport from './Transport.js'
 
 describe('Mppx', () => {
+  test('exports managed session and low-level session method separately', () => {
+    expectTypeOf(session({ account: {} as Account })).toHaveProperty('fetch')
+    expectTypeOf(sessionMethod({ account: {} as Account })).toMatchTypeOf<Method.AnyClient>()
+  })
+
   test('has methods array', () => {
     const method = charge({
       account: {} as Account,
